Separate store and slice
