+++ app/javascript/application.js
@@ -4,7 +4,7 @@ import "./controllers"
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { store } from './store';
+import store from './store';
 import Greeting from './greeting';
 
 ReactDOM.render(
