+++ app/javascript/greeting.js
@@ -1,9 +1,10 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { fetchGreeting } from './store';
+import { fetchGreeting } from './greetingSlice';
 
 function Greeting() {
   const dispatch = useDispatch();
+
   const greeting = useSelector(state => state.greeting.greeting);
   const loading = useSelector(state => state.greeting.loading);
   const error = useSelector(state => state.greeting.error);
